+++ src/three/plugins/index.js
@@ -16,6 +16,8 @@ export * from './mvt/MVTAnnotationsPlugin.js';
 
 // other formats
 export * from './potree/PotreePlugin.js';
+export * from './pointcloud/PointCloudEffectsPlugin.js';
+export * from './pointcloud/PointCloudMaterial.js';
 export * from './images/GeneratedSurfacePlugin.js';
 export * from './images/terrain-rgb/TerrainRGBMeshPlugin.js';
 export * from './images/terrain-rgb/TerrariumMeshPlugin.js';
